♻️ refactor: MakeDoucmentButton 컴포넌트 분리 #24
